fix unit_tests
